requirements.txt
beautifulsoup4 (4.6.0)
bleach (2.0.0)
certifi (2017.7.27.1)
chardet (3.0.4)
click (6.7)
colorama (0.3.9)
configparser (3.5.0)
decorator (4.1.2)
Django (1.11.4)
entrypoints (0.2.3)
et-xmlfile (1.0.1)
Flask (0.12.2)
html5lib (0.999999999)
idna (2.6)
ipykernel (4.6.1)
ipython (6.1.0)
ipython-genutils (0.2.0)
ipywidgets (7.0.0)
itsdangerous (0.24)
jdcal (1.3)
jedi (0.10.2)
Jinja2 (2.9.6)
jsonschema (2.6.0)
jupyter (1.0.0)
jupyter-client (5.1.0)
jupyter-console (5.2.0)
jupyter-core (4.3.0)
kafka (1.3.4.1)
lxml (3.8.0)
MarkupSafe (1.0)
mistune (0.7.4)
mysql-connector (2.1.6)
mysqlclient (1.3.10)
nbconvert (5.3.1)
nbformat (4.4.0)
notebook (5.0.0)
olefile (0.44)
openpyxl (2.4.8)
pandocfilters (1.4.2)
pdfminer.six (20170720)
pdfminer3k (1.3.1)
pickleshare (0.7.4)
Pillow (4.2.1)
pip (9.0.1)
ply (3.10)
prompt-toolkit (1.0.15)
py (1.4.34)
pycryptodome (3.4.6)
Pygments (2.2.0)
PyPDF2 (1.26.0)
pytesseract (0.1.7)
pytest (3.2.1)
python-dateutil (2.6.1)
python-mimeparse (1.6.0)
pytz (2017.2)
PyYAML (3.12)
pyzmq (16.0.2)
qtconsole (4.3.1)
requests (2.18.4)
setuptools (28.8.0)
simplegeneric (0.8.1)
six (1.10.0)
SQLAlchemy (1.1.14)
testpath (0.3.1)
tornado (4.5.2)
traitlets (4.3.2)
urllib3 (1.22)
virtualenv (15.1.0)
wcwidth (0.1.7)
webencodings (0.5.1)
Werkzeug (0.12.2)
widgetsnbextension (3.0.2)
win-unicode-console (0.5)